Move reading related tools from util.aws to reading.util
